% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_2ch_menu}
\alias{get_2ch_menu}
\title{Get 2ch menu from 'bbsmenu.html'}
\usage{
get_2ch_menu(server_url, encoding = "Shift_JIS")
}
\arguments{
\item{server_url}{A string.}

\item{encoding}{A string.}
}
\value{
The menu object.
}
\description{
Get 2ch menu from 'bbsmenu.html'
}
\examples{
get_2ch_menu("http://2ch.sc")
}
